requirements.txt
boto3
requests
ipython
ipdb
sentence-transformers
faiss-cpu
